+++ lib/scss_lint/utils.rb
@@ -34,6 +34,16 @@ module SCSSLint
         hex[5] == hex[6]
     end
 
+    # Takes a string like `hello "world" 'how are' you` and turns it into:
+    # `hello   you`.
+    # This is useful for scanning for keywords in shorthand properties or lists
+    # which can contain quoted strings but for which you don't want to inspect
+    # quoted strings (e.g. you care about the actual color keyword `red`, not
+    # the string "red").
+    def remove_quoted_strings(string)
+      string.gsub(/"[^"]*"|'[^']*'/, '')
+    end
+
   private
 
     INVALID_NAME_CHARS = '[_A-Z]'